Changed key from Long to Integer

+++ ToreadorRESTAPI/src/main/java/uk/ac/city/toreador/rest/api/jpa/repositories/CompositeServicesRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import uk.ac.city.toreador.rest.api.entities.CompositeService;
 
-public interface CompositeServicesRepository extends CrudRepository<CompositeService, Long>{
+public interface CompositeServicesRepository extends CrudRepository<CompositeService, Integer>{
 	
 }
